+++ src/pages/Home/Home.jsx
@@ -4,12 +4,14 @@ import { connect } from "react-redux";
 import TabsRovers from "../../components/tabsRover/TabsRovers";
 import CardPhoto from "../../components/cardPhoto/CardPhoto";
 import Paginate from "../../components/paginate/Paginate";
-import Skeleton from '../../components/skeleton/Skeleton'
+import Skeleton from '../../components/skeleton/Skeleton';
+
+import Swal from 'sweetalert2';
 
 const Home = ({ STATE, CONTENT, LOADING }) => {
   
   const [currentPage, setCurrentPage] = useState(1);
-  const [postPerPage] = useState(25);
+  const [postPerPage] = useState(24);
 
   const indexOfLastPage = currentPage * postPerPage;
   const indexOfFirstPage = indexOfLastPage - postPerPage;
@@ -31,31 +33,83 @@ const Home = ({ STATE, CONTENT, LOADING }) => {
 
       let storage = localStorage.getItem("photos")
       if(storage){
-          let previous = JSON.parse(storage);
-          let arr = []
+        Swal.fire({
+          title: 'Do you want to bookmark this photo?',
+          text: "it's a nice photo!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, add it!'
+        }).then((result) => {
+          
+          if (result.isConfirmed) {
+            let previous = JSON.parse(storage);
+            let arr = []
 
-          previous.forEach(element => {
-            if(element.id !== photo.id){
-              arr.push(element)
-              localStorage.clear()
-              localStorage.setItem("photos",JSON.stringify(arr.concat(photo))) 
-            }else{
-              alert('esta foto ya existe en tus favoritos')
-            }
-          })     
+            previous.forEach(element => {
+              if(element.id !== photo.id){
+                arr.push(element)
+                localStorage.clear()
+                localStorage.setItem("photos",JSON.stringify(arr.concat(photo))) 
+                Swal.fire(
+                  'Added!',
+                  'Your photo is now in the favorites section',
+                  'success'
+                )
+              }else{
+                Swal.fire(
+                  'Oh! Oh!',
+                  'this photo already exists in your section',
+                  'info'
+                  )
+              }
+            })     
+          }else if(result.dismiss || result.isDenied){
+            Swal.fire(
+              'Good idea!',
+              'There are better photos!',
+              'info'
+            )
+          }
+        })
           return
       }else{
-        localStorage.setItem("photos",JSON.stringify([photo]))
+        Swal.fire({
+          title: 'Do you want to bookmark this photo?',
+          text: "it's a nice photo!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, added it!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            localStorage.setItem("photos",JSON.stringify([photo]))
+            Swal.fire(
+              'Added!',
+              'Your photo is now in the favorites section.',
+              'success'
+            )
+          }else if(result.dismiss || result.isDenied){
+            Swal.fire(
+              'Good idea!',
+              'There are better photos!',
+              'info'
+            )
+          }
+        })
+        
       }
   }
 
   return (
     <div className="container_home">
+      <h1>Mars Rover Photos</h1>
       <TabsRovers />
       <div>
         {STATE && !LOADING ? (
           <>
-          <h2>photos of the day</h2>
             <div className="container_cards">
               {currentPost.map((elem, i) => (
                 <CardPhoto props={elem} key={i} addFav={addFavorite} removeFav={false}/>
